Move kinvey package
