--- packages/kinvey/package.json
@@ -1,26 +1,0 @@
-{
-  "private": true,
-  "version": "3.9.3-beta.7",
-  "name": "kinvey",
-  "main": "src/index.js",
-  "dependencies": {
-    "es6-promise": "4.1.1",
-    "kinvey-client": "^3.9.3-beta.7",
-    "kinvey-errors": "^3.9.3-beta.7",
-    "kinvey-request": "^3.9.3-beta.7",
-    "kinvey-user": "^3.9.3-beta.7",
-    "kinvey-utils": "^3.9.3-beta.7"
-  },
-  "devDependencies": {
-    "chai": "4.1.2",
-    "chai-as-promised": "7.1.1",
-    "mocha": "4.0.1",
-    "nock": "9.0.28",
-    "request": "2.83.0",
-    "sinon": "4.0.1",
-    "sinon-chai": "2.14.0"
-  },
-  "scripts": {
-    "test": "mocha -s 250 test/**/*.spec.js"
-  }
-}
